+++ frontend/modules/profiles/actions/register.php
@@ -110,15 +110,16 @@ class FrontendProfilesRegister extends FrontendBaseBlock
 			// no errors
 			if($this->frm->isCorrect())
 			{
-				// generate salt
-				$salt = FrontendProfilesModel::getRandomString();
-
 				// init values
+				$settings = array();
 				$values = array();
+
+				// generate salt
+				$settings['salt'] = FrontendProfilesModel::getRandomString();
 
 				// values
 				$values['email'] = $txtEmail->getValue();
-				$values['password'] = FrontendProfilesModel::getEncryptedString($txtPassword->getValue(), $salt);
+				$values['password'] = FrontendProfilesModel::getEncryptedString($txtPassword->getValue(), $settings['salt']);
 				$values['status'] = 'inactive';
 				$values['display_name'] = $txtEmail->getValue();
 				$values['registered_on'] = FrontendModel::getUTCDate();
@@ -139,17 +140,16 @@ class FrontendProfilesRegister extends FrontendBaseBlock
 					FrontendModel::triggerEvent('profiles', 'after_register', array('id' => $profileId));
 
 					// generate activation key
-					$activationKey = FrontendProfilesModel::getEncryptedString($profileId . microtime(), $salt);
+					$settings['activation_key'] = FrontendProfilesModel::getEncryptedString($profileId . microtime(), $settings['salt']);
 
 					// set settings
-					FrontendProfilesModel::setSetting($profileId, 'salt', $salt);
-					FrontendProfilesModel::setSetting($profileId, 'activation_key', $activationKey);
+					FrontendProfilesModel::setSettings($profileId, $settings);
 
 					// login
 					FrontendProfilesAuthentication::login($profileId);
 
 					// activation URL
-					$mailValues['activationUrl'] = SITE_URL . FrontendNavigation::getURLForBlock('profiles', 'activate') . '/' . $activationKey;
+					$mailValues['activationUrl'] = SITE_URL . FrontendNavigation::getURLForBlock('profiles', 'activate') . '/' . $settings['activation_key'];
 
 					// send email
 					FrontendMailer::addEmail(