+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Improvements:
 
 * Core: Upgraded to CKEditor 3.6.4
+* Profiles: rewrote method for inserting/multiple settings.
 
 Bugfixes: